requirements.txt
flask
pillow
numpy
opencv-python
gunicorn
